Remove `use_frameworks` (#58)
`use_frameworks` seems not to be needed. So I've removed that call and built the app - everything seems to work ;) 

We should avoid using frameworks as long as possible. This option will add another step that users have to apply during the installation process. Also, it won't work great with the expo ecosystem. In theory, we added support for frameworks some time ago, but I don't trust that change 😅