Camera Animations for Google and Apple Maps (#93)
# Camera Move

## Functionality

Type animation is done by accessing the map `ref` and calling `ref.current.moveCamera(CameraMove)` . The `CameraMove` type consists of values:

- `target`
- `bearing`
- `tilt`
- `zoom`
- `latLngDelta`
- `duration`
- `animate`

The type is similar to `CameraPosition`, but all fields except for `duration`  and `animate` are optional. If any of the fields is left as null the camera will keep the current value of that position parameter. When both `latLngDelta` is set the `bearing`, `tilt`, `zoom` fields will be ignored, and the camera will move to smallest view containing a bounding box created around the `target` by `latLngDelta.latitudeDelta` and `latLngDelta.longitudeDelta` (target is in the center of the box). 

Due to platform differences `zoom` might provide a  slightly different view for Apple Maps and Google Maps, since a conversion is needed from Google Maps’ `zoom` to Apple Maps’ `altitude`